requirements.txt
# Edge-Cloud Code Analysis Prototype - Dependencies

# Core async and HTTP（asyncio 为 Python 3 标准库，勿安装 PyPI 的 asyncio 包）
aiohttp==3.9.1
httpx==0.25.2

# AST parsing
tree-sitter==0.21.0
tree-sitter-python==0.21.0
tree-sitter-java==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-cpp==0.21.0

# Cloud LLM clients（Anthropic 等走 OpenAI 兼容 base_url 时仅需 openai 包）
openai==1.6.1

# Data validation and schemas
pydantic==2.5.3
pydantic-settings==2.1.0

# Configuration
pyyaml==6.0.1

# Logging
structlog==23.2.0
colorama==0.4.6

# No database (budget is in-memory)

# CLI
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Code quality
black==23.12.1
ruff==0.1.8
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Clone detection eval: UniXcoder fine-tuned classifier (evaluate_clone_detection.py)
transformers>=4.36.0
datasets>=2.14.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
numpy>=1.24.0
# Cross-Encoder 精排（evaluate_code_search / csn_ce_rerank）
sentence-transformers>=3.0.0
